Execute Shutdown Command: Skip if no POWER_CTRL

+++ keywords.robot
@@ -799,11 +799,19 @@ Restore Initial DUT Connection Method
     END
 
 Execute Shutdown Command
-    [Documentation]    Windows shutdown keyword, checks power LED state where available.
+    [Documentation]    Shutdown keyword, checks power LED state where available.
     ...    Depends on existing SSH connection to DUT, restores initial connection method
     ...    after power loss.
-    VAR    ${DUT_CONNECTION_METHOD}=    SSH    scope=GLOBAL
-    Execute Command In Terminal    shutdown /s /f /t 0
+    IF    '${POWER_CTRL}' == 'none'
+        # Skip if there is no way to boot the device back up after shutdown.
+        RETURN
+    END
+    IF    '${BOOTED_OS_ID}'.startswith("2")
+        Execute Command In Terminal    shutdown 0
+    ELSE IF    '${BOOTED_OS_ID}' == '${ENV_ID_WINDOWS}'
+        VAR    ${DUT_CONNECTION_METHOD}=    SSH    scope=GLOBAL
+        Execute Command In Terminal    shutdown /s /f /t 0
+    END
     IF    '${CHECK_POWER_LED_SUPPORT}' == '${TRUE}'
         ${loop_iterations}=    Evaluate    ${WINDOWS_SHUTDOWN_AWAITING_SECONDS} * 2
         FOR    ${i}    IN RANGE    ${loop_iterations}
